+++ packages/template-typescript/tailwind.config.js
@@ -1,0 +1,3 @@
+// We create a dummy tailwind configuration file so that tooling like the vscode extension
+// provide syntax hints.
+module.exports = {};