+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for picking among first-party `create-nostalgie-app` templates. With this release, we include `javascript` and `typescript`.
+
+- Added a simple (but fun) Server Function to the default templates. This function demonstrates using the `useQueryFunction` hook with a Server Function. The hook is configured to keep the initial, server-rendered value for 5s and then cycle every 5s.
+
+### Changed
+- Changed default `create-nostalgie-app` to use the JavaScript template.
 
 ## [0.70.0] - 2021-02-11
 ### Changed